Move sensor constants and define public names from this module
